+++ src/xrGame/Entity.h
@@ -110,7 +110,7 @@ public:
 
 	virtual void Die(CObject* who);
 	//			void			KillEntity			(CObject* who);
-	void KillEntity(u16 whoID, BOOL bypass_actor_check = FALSE);
+	void KillEntity(u32 whoID, BOOL bypass_actor_check = FALSE);
 
 	// Events
 	virtual void OnEvent(NET_Packet& P, u16 type);
@@ -132,7 +132,7 @@ private:
 	ALife::_OBJECT_ID m_killer_id;
 
 public:
-	IC u16 killer_id() const { return m_killer_id; };
+	IC u32 killer_id() const { return m_killer_id; };
 	virtual bool use_simplified_visual() const { return false; };
 
 public:
